refactor: make namespaces correct
[skip changelog]
