+++ Libplanet.Tests/PoS/DelegateCtrlTest.cs
@@ -1,6 +1,8 @@
 using Libplanet.Action;
 using Libplanet.Crypto;
 using Libplanet.PoS;
+using Libplanet.PoS.Control;
+using Libplanet.PoS.Model;
 using Xunit;
 
 namespace Libplanet.Tests.PoS
