+++ Libplanet.Tests/PoS/ValidatorTest.cs
@@ -1,5 +1,6 @@
 using Libplanet.Crypto;
 using Libplanet.PoS;
+using Libplanet.PoS.Model;
 using Xunit;
 
 namespace Libplanet.Tests.PoS
