+++ Libplanet/PoS/Action/Sys/Redelegate.cs
@@ -1,6 +1,7 @@
 using Bencodex.Types;
 using Libplanet.Assets;
 using Libplanet.PoS;
+using Libplanet.PoS.Control;
 
 namespace Libplanet.Action.Sys
 {
