+++ Libplanet/PoS/Control/Bond.cs
@@ -1,7 +1,8 @@
 using Libplanet.Action;
 using Libplanet.Assets;
+using Libplanet.PoS.Model;
 
-namespace Libplanet.PoS
+namespace Libplanet.PoS.Control
 {
     internal static class Bond
     {
