+++ Libplanet/PoS/Control/DelegateCtrl.cs
@@ -1,7 +1,7 @@
 using Libplanet.Action;
 using Libplanet.Assets;
 
-namespace Libplanet.PoS
+namespace Libplanet.PoS.Control
 {
     internal static class DelegateCtrl
     {
