+++ Libplanet/PoS/Control/ValidatorPowerIndexCtrl.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using Libplanet.Action;
 using Libplanet.Assets;
 
-namespace Libplanet.PoS
+namespace Libplanet.PoS.Control
 {
     internal static class ValidatorPowerIndexCtrl
     {
